requirements.txt
absl-py==0.13.0
aiofiles==0.7.0
appnope==0.1.2
asgiref==3.4.1
astor==0.8.1
astunparse==1.6.3
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.1
click==8.0.1
decorator==5.0.9
fastapi==0.66.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h11==0.12.0
h5py==3.1.0
idna==3.2
imageio==2.9.0
importlib-metadata==4.6.1
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
matplotlib-inline==0.1.2
numpy==1.19.5
oauthlib==3.1.1
opencv-contrib-python==4.5.3.56
opt-einsum==3.3.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
Pygments==2.9.0
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
starlette==0.14.2
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-docs @ git+https://github.com/tensorflow/docs@47a807024a49d648ff635434d49e3f03a8cfc57a
tensorflow-estimator==2.5.0
tensorflow-hub==0.12.0
termcolor==1.1.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.6
uvicorn==0.14.0
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
